requirements.txt
# API and Framework
fastapi==0.112.2
uvicorn==0.30.6
python-multipart==0.0.19
aiofiles==24.1.0
jinja2==3.1.6

# Auth / Security
bcrypt==4.0.1
cryptography==44.0.1
python-jose==3.4.0
passlib==1.7.4

# Database
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
alembic==1.13.2

# AI (Gemini + LangChain)
google-generativeai==0.7.2
langchain==0.2.15
langchain-google-genai==1.0.10

# Helpers
python-dotenv==1.0.1
markdown==3.7
bs4==0.0.2

# Test
pytest==8.3.2
pytest-asyncio==0.24.0
httpx==0.27.2
